Use longs instead

+++ src/main/scala/sqid/Sqids.scala
@@ -13,13 +13,13 @@ import sqids.options.InvalidSqidsOptions
 import sqids.options.SqidsOptions
 
 trait Sqids {
-  def encodeUnsafeString(numbers: Int*): String
-  def encodeUnsafe(numbers: Int*): Sqid
-  def encode(numbers: Int*): Either[SqidsError, Sqid]
-  def encode(numbers: List[Int]): Either[SqidsError, Sqid]
-  def decode(id: String): List[Int]
-  def minValue: Int
-  def maxValue: Int
+  def encodeUnsafeString(numbers: Long*): String
+  def encodeUnsafe(numbers: Long*): Sqid
+  def encode(numbers: Long*): Either[SqidsError, Sqid]
+  def encode(numbers: List[Long]): Either[SqidsError, Sqid]
+  def decode(id: String): List[Long]
+  def minValue: Long
+  def maxValue: Long
 }
 
 object Sqids {
@@ -41,40 +41,40 @@ object Sqids {
     val _alphabet = options.alphabet.shuffle
     new Sqids {
 
-      override def encodeUnsafe(numbers: Int*): Sqid = encode(numbers: _*) match {
+      override def encodeUnsafe(numbers: Long*): Sqid = encode(numbers: _*) match {
         case Left(value) => throw value
         case Right(value) => value
       }
 
-      override def encodeUnsafeString(numbers: Int*): String = encode(numbers: _*) match {
+      override def encodeUnsafeString(numbers: Long*): String = encode(numbers: _*) match {
         case Left(error) => throw error
         case Right(value) => value.value
       }
 
-      override def encode(numbers: Int*): Either[SqidsError, Sqid] =
+      override def encode(numbers: Long*): Either[SqidsError, Sqid] =
         encode(numbers.toList)
 
-      override def encode(numbers: List[Int]): Either[SqidsError, Sqid] =
+      override def encode(numbers: List[Long]): Either[SqidsError, Sqid] =
         encode_(numbers)
 
-      override def minValue: Int = 0
+      override def minValue: Long = 0
 
-      override def maxValue: Int = Int.MaxValue
+      override def maxValue: Long = Long.MaxValue
 
-      override def decode(input: String): List[Int] =
+      override def decode(input: String): List[Long] =
         input match {
           case "" => List.empty
           case id if !_alphabet.validId(id) => List.empty
           case id => getNumbers(id.head, id.tail)
         }
 
-      private def getNumbers(prefix: Char, id: String): List[Int] = {
+      private def getNumbers(prefix: Char, id: String): List[Long] = {
         @tailrec
         def go(
           id: String,
           alphabet: Alphabet,
-          acc: Vector[Int]
-        ): List[Int] =
+          acc: Vector[Long]
+        ): List[Long] =
           if (id.isEmpty) acc.toList
           else
             alphabet.splitAtSeparator(id) match {
@@ -97,7 +97,7 @@ object Sqids {
       }
 
       private def encode_(
-        numbers: List[Int]
+        numbers: List[Long]
       ): Either[SqidsError, Sqid] =
         numbers match {
           case numbers if numbers.exists(i => i > maxValue || i < minValue) =>
